+++ dev_kit_mcp_server/tools/core/__init__.py
@@ -1,9 +1,7 @@
 """core repo operations."""
 
-from dev_kit_mcp_server.tools.core.file_ops import AsyncOperation, FileOperation, _Operation
+from dev_kit_mcp_server.tools.core.file_base import _Operation
+from dev_kit_mcp_server.tools.core.ops_async import AsyncOperation
+from dev_kit_mcp_server.tools.core.ops_sync import FileOperation
 
-__all__ = [
-    "FileOperation",
-    "AsyncOperation",
-    "_Operation",
-]
+__all__ = ["AsyncOperation", "_Operation", "FileOperation"]
